add and delete functions moved to delete-completed-tasks for easiesr assesment
