+++ package.json
@@ -29,6 +29,7 @@
     "eslint-plugin-import": "^2.27.5",
     "hint": "^7.1.8",
     "html-webpack-plugin": "^5.5.1",
+    "jest": "^29.5.0",
     "lighthouse": "^10.2.0",
     "style-loader": "^3.3.2",
     "stylelint": "^13.13.1",
