+++ src/addDeleteEvent.js
@@ -1,16 +1,11 @@
-import Task from './Task.js';
+import { addNewTask } from "./delete-completed-tasks";
 const eventListeners = (newList) => {
   const newChore = document.querySelector('.addChore');
   const deleteButonn = document.querySelector('.clearBtnn');
   
   newChore.addEventListener('keydown', (event) => {
     if (event.key === 'Enter') {
-      const description = newChore.value.trim();
-      if (description !== '') {
-        const task = new Task(description, false, newList.length + 1);
-        newList.addTask(task);
-        newChore.value = '';
-      }
+      addNewTask(newChore, newList);
     }
   });
